Refactor : step1 1차 피드백반영
- 코드네이밍 수정
- init 삭제
